Reorganizing the project
